fixed larastan warning

+++ app/Http/Controllers/Ecommerce/Payments/CashOnDeliveryController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers\Ecommerce\Payments;
 
 use App\Http\Controllers\Controller;
 use App\Mail\OrderPlacedMail;
+use App\Models\Cart;
 use Illuminate\Http\Request;
 use App\Models\Order;
 use App\Models\OrderItem;
+use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Mail;
@@ -15,11 +17,11 @@ class CashOnDeliveryController extends Controller
 {
     public function cashPaymentProcess(Request $request): RedirectResponse
     {
-        $user=auth()->user();
+        $user=User::findOrFail(auth()->id());
 
         $order=Order::create([
             "user_id"=>$user->id,
-            "delivery_information_id"=>$user->deliveryInformation->id,
+            "delivery_information_id"=>$user->deliveryInformation ? $user->deliveryInformation->id : null,
             'payment_type'=>"cash on delivery",
             'payment_method'=>"cash on delivery",
             'total_amount'=>$request->total_price,
@@ -46,19 +48,26 @@ class CashOnDeliveryController extends Controller
 
         $placedOrder=Order::with(["deliveryInformation","orderItems.product.shop"])->where("id", $order->id)->first();
 
+
         Mail::to($placedOrder->deliveryInformation->email)->send(new OrderPlacedMail($placedOrder));
 
         if (session("coupon")) {
             session()->forget("coupon");
         }
 
-        $cart=auth()->user()->cart;
-        $cartItems=$cart->cartItems;
+        $cart=Cart::where("user_id", auth()->id())->first();
 
-        $cartItems->each(function ($item) {
-            $item->destroy($item->id);
-        });
+        if($cart) {
 
+            $cartItems=$cart->cartItems;
+
+            $cartItems->each(function ($item) {
+
+                $item->destroy($item->id);
+
+            });
+
+        }
 
         return to_route("home")->with("success", "Your place order is successfully");
     }
